Removal of Marisa for the time being
It is unclear whether she recreationally engages in these sapphic relationships or is just genuinely this downbad. In the latter case she would be the ?intensified synonym? of a Yukari.

On an unrelated note, I think it would be nice to have a single grammatical term for nouns that represent a stronger or more intense form of another noun. This would be similar to a superlative for adverbs.